requirements.txt
# Django core
Django>=5.0,<6.0

# Django REST Framework
djangorestframework>=3.15

# JWT authentication
djangorestframework-simplejwt>=5.3

# OpenAPI / Swagger
drf-spectacular>=0.27

# Image processing
Pillow>=10.0

# Environment variables
python-dotenv>=1.0

# Production server (opcional pero recomendado)
gunicorn>=21.2

Pillow==10.3.0

psycopg2-binary==2.9.9
